requirements.txt
Django==4.2.1
django-appconf==1.0.5
django-browser-reload==1.8.0
django-ckeditor==6.6.1
django-classy-tags==4.0.0
django-cms==3.11.3
django-crispy-forms==2.0
django-filer==2.2.4
django-formtools==2.4.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-tailwind==3.5.0
django-treebeard==4.7
djangocms-admin-style==3.2.4
djangocms-attributes-field==3.0.0
djangocms-revealjs==0.3
djangocms-text-ckeditor==5.1.2
requests==2.30.0
whitenoise==6.4.0
django-tinymce
